@@ -13,10 +13,11 @@ if (process.argv.length !== 5 && process.argv.length !== 6) {
 	process.exit(1);
 }
 
-rmdirRecursive(typingsPackageName);
+rmdirRecursive(path.join("types", typingsPackageName));
 const notNeededPackages = JSON.parse(fs.readFileSync("notNeededPackages.json", "utf-8"));
 notNeededPackages.packages.push({ libraryName, typingsPackageName, sourceRepoURL, asOfVersion });
-fs.writeFileSync("notNeededPackages.json", JSON.stringify(notNeededPackages, undefined, 4), "utf-8");
+notNeededPackages.packages.sort((x, y) => x.typingsPackageName < y.typingsPackageName ? -1 : 1);
+fs.writeFileSync("notNeededPackages.json", JSON.stringify(notNeededPackages, undefined, 4) + "\n", "utf-8");
 
 function rmdirRecursive(dir) {
 	for (let entry of fs.readdirSync(dir)) {
